remove direct call to methods::`slot<-`
`slot<-` from methods package is a wrapper around R_do_slot_assign
function from GNU R. This function is being called with the assumption
that assignments to slots are performed in-place and thus no new copies
are being made. This is not true, as R_do_slot_assign will make new
copies if the named bit is set (object is shared), see this line in GNU
R C code:
https://github.com/wch/r-source/blob/d00ce14c27db627d8644cce8551b792362cc46bf/src/main/attrib.c#L1784

Therefore this additional steps will not give any improvements over call
to slot()<- and depends on implementation details, changes to which
could break this function and packages dependending on this (eg
SummarizedExperiment rbind function).